Removing articlesLayout route Thes component does not qualify as route and gives build time error. I believe mistakenly this file is being created inside of pages folder instead of component folder. So, removing from pages folder and putting inside component folder, in case it is needed later.
